backend/requirements.txt
pip>=20.0.2
Markdown>=3.1.1
gunicorn>=20.0.4
json-logging>=1.0.5
psycopg2-binary>=2.8.4

# django
Django == 3.0.6
django-model-utils>=4.0.0
django-extra-fields>=2.0.2
django-storages>=1.8
django-filter>=2.2.0
django-cors-headers>=2.4.0
django-rest-email-auth

# REST API
djangorestframework>=3.11.0
elasticsearch-dsl==7.1.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf>=0.20.5

# analysis
ipykernel>=5.1.3
pandas>=0.25.3
numpy>=1.18.1
scipy>=1.4.1
matplotlib>=3.1.2
pint>=0.10-1

-e git://github.com/matthiaskoenig/pkdb_analysis.git@bf6077a4702838886f148c4e70de0b7ec2a3f3b9#egg=pkdb_analysis

